requirements.txt
gunicorn==19.7.1
python-dotenv==0.21.1
requests
flask_sqlalchemy
flask_admin
flask_login
flask_migrate
datetime
pandas
openai
lxml==4.7.1

